azure_function/requirements.txt
azure-functions
azure-storage-blob
openai
python-docx
PyMuPDF
python-dotenv
